Update cmake config
