+++ CMakeLists.txt
@@ -1,23 +1,18 @@
 cmake_minimum_required(VERSION 3.0.0)
 project(GYM_APP VERSION 1.0.0)
 
-include(CTest)
-enable_testing()
-
-file(GLOB src CONFIGURE_DEPENDS
-    "./src/*.h"
-    "./src/.hpp"
-    "./src/*.cpp"
-    "./src/*.c"
-)
+include(GNUInstallDirs)
+set(CMAKE_ARCHIVE_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/${CMAKE_INSTALL_LIBDIR})
+set(CMAKE_LIBRARY_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/${CMAKE_INSTALL_LIBDIR})
+set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/${CMAKE_INSTALL_BINDIR})
 
 add_subdirectory(googletest)
 include_directories(googletest/googletest/include)
 
-add_executable(app ${src})
-add_subdirectory(tests)
+add_subdirectory(src)
 
-set_property(TARGET app PROPERTY CXX_STANDARD 17)
+enable_testing()
+add_subdirectory(tests)
 
 set(CPACK_PROJECT_NAME ${PROJECT_NAME})
 set(CPACK_PROJECT_VERSION ${PROJECT_VERSION})
