+++ src/includes/invoice.h
@@ -4,5 +4,4 @@
 #include <string>
 #include <vector>
 
-std::string currentDateTime();
 void generateInvoice(std::map<std::string, double> &details);
